fix Apache Maven POM for SLF4j badness
The whole point of the slf4j facade is to allow libraries
to NOT provide a logging implementation. When we provide
a hard coded logging implementation here, in this case,
slf4j-log4j12, we are doing the exact same thing that
is a problem with Spring Framework, e.g.

This forces everyone to exclude slf4j-log412 when they
do not want to use Log4j as the final logging
implementation.

Please do not continue this bad behavior.