Using default dozer configuration is not worth logging a warning
If you want to monitor a log file of an application in production you would define this basic rules:
ERROR: Severe problem. Immediately get a person of charge out of his bed and let him fix the problem.
WARNING: Write an email to the responsible person and let him fix the problem within a day.

Because the information "using defaults" is not really a production problem, I would recommend to only log at information level.